requirements.txt
azureml-core
azureml-dataprep
azureml-train-core
pandas
torch
torchvision
tqdm
Unidecode